환경 효과 저장 안되는 문제 수정

+++ Assets/Scripts/Manager/Data_Manager.cs
@@ -742,7 +742,7 @@ public class Data_Manager : MonoBehaviour
         if (Sound_Effect == value) return;
         
         Sound_Effect = value;
-        PlayerPrefs.SetFloat("Sound_Effect", value);
+        PlayerPrefs.SetInt("Sound_Effect", value);
     }
 
     public void SetSound_UI(int value)
